Add drag-and-drop queue reordering, closes #34 (#80)
The easiest way to implement this is to harness GtkTreeView's
reorderable property. However, this doesn't work with
GtkTreeModelFilter, so the TreeView's model gets switched based on
whether filtering is active.

To facilitate reordering, I added playlist position as a column, so when
the row data gets inserted in the new position, the TreePath (row index)
and pos column can be used as move target and source, respectively.

After issuing the move command, the playlist status will change and
update the queue with the new positions.

yktoo: This is an initial implementation, limited to dragging a single row at a time. This can be improved in the future by implementing multi-row drag-n-drop.